requirements2.txt
absl-py==0.15.0
ansicon==1.89.0
appdirs==1.4.4
blessed==1.19.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.2.0
cycler==0.11.0
docker-pycreds==0.4.0
flyingcircus==0.1.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna==3.3
ImageHash==4.2.1
imageio==2.22.1
jinxed==1.2.0
joblib==1.2.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
networkx==2.8.7
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.4
pathtools==0.1.2
piecewise-distortion==0.1
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.4
python-dateutil==2.8.2
pytk==0.0.2.1
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.7.2
seaborn==0.11.2
sentry-sdk==1.5.0
setuptools-scm==7.0.5
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.7
wcwidth==0.2.5
Werkzeug==2.0.2
wincertstore==0.2
yaspin==2.1.0
